drm/rockchip: dsi: add ability to work as a phy instead of full dsi
SoCs like the rk3288 and rk3399 have 3 mipi dphys on them. One is TX-only,
one is RX-only and one can be configured to do either TX or RX.

The RX dphy is controlled externally through registers in the "General
Register Files", while the other two are controlled through the
"Configuration and Test Interface" inside the DSI controller io-memory
area.

The Rockchip dw-dsi controller already controls these dphys for the
TX case in the driver, but when we want to also allow configuration
for RX to the ISP from the media subsystem we need to expose phy-
functionality instead.

So add a bit of infrastructure to allow the dsi driver to work as a phy
and make sure it can be only one or the other at a time.

Similarly as the dsi-controller will be part of the drm-graph when active,
add an empty component to the drm-graph when in phy-mode.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>