arm64: dts: Add the OV13850 and OV4689 to dts
Add the two offical friendlyElec cameras to the device tree, while
activating the second instance of the rkisp1 and connect it with the
mipi_dsi1, which is used as a dphy.
Mostly a copy of what is found on the friendlyElec fork of the rockchip
downstream BSP kernel tree.
Additionally, increase the temperature limits on the CPU.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>